Fixes issue on handling the strength of a token
`TokenGenerator#token` would take a value and divide it by 2. This would
cause an issue where odd numbers being passed would be returned as a float,
obviously. However, this presented an issue.

As a token is generated with `random_bytes` wrapped inside `bin2hex`,
passing the strength of `32` would generate a string length of
`64`. This would cause another issue where if an odd number was given for
the strength, it would give the token value length of the original odd
value + 1.

This will fix that issue and only return a token value length that is
equal to the original desired strength value.